refactor: wrapper타입 변수를 primitive 타입으로 변환

+++ src/main/java/lotto/domain/LottoGroup.java
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class LottoGroup {
-    public static final long LOTTO_PRICE = 1000;
+    public static final int LOTTO_PRICE = 1000;
     private final List<Lotto> lottos;
 
     public LottoGroup(final List<Lotto> lottos) {
